+++ pandapower/__init__.py
@@ -17,5 +17,6 @@ from pandapower.powerflow import *
 from pandapower.opf import *
 from pandapower.optimal_powerflow import OPFNotConverged
 from pandapower.pf.runpp_3ph import runpp_3ph
+from pandapower.groups import *
 import pandas as pd
 pd.options.mode.chained_assignment = None  # default='warn'