+++ pandapower/test/converter/test_from_ppc.py
@@ -39,18 +39,16 @@ max_diff_values1 = {"bus_vm_pu": 1e-6, "bus_va_degree": 1e-5, "branch_p_mw": 1e-
                     "branch_q_mvar": 1e-3, "gen_p_mw": 1e-3, "gen_q_mvar": 1e-3}
 
 
-def get_testgrids(name, filename):
+def get_testgrids(foldername, filename):
     """
     This function return the ppc (or pp net) which is saved in ppc_testgrids.p to validate the
     from_ppc function via validate_from_ppc.
     """
-    folder = os.path.join(pp.pp_dir, 'test', 'converter')
+    folder = os.path.join(pp.pp_dir, 'test', 'converter', foldername)
     file = os.path.join(folder, filename)
-    if filename.endswith(".json"):
-        ppcs = pp.from_json(file)
-    elif filename.endswith(".p"):
-        ppcs = pickle.load(open(file, "rb"))
-    return ppcs[name]
+    convert = "ppc" in filename
+    ppcs = pp.from_json(file, convert=convert)
+    return ppcs
 
 
 def validate_other_than_py37(ppc, net, max_diff_values):
@@ -62,7 +60,7 @@ def validate_other_than_py37(ppc, net, max_diff_values):
 
 
 def test_from_ppc_simple_against_target():
-    ppc = get_testgrids('case2_2', 'ppc_testgrids.json')
+    ppc = get_testgrids('ppc_testgrids', 'case2_2.json')
     net_by_ppc = from_ppc(ppc)
     net_by_code = pp.from_json(os.path.join(pp.pp_dir, 'test', 'converter', 'case2_2_by_code.json'))
     pp.set_user_pf_options(net_by_code)  # for assertion of nets_equal
@@ -78,7 +76,7 @@ def test_from_ppc_simple_against_target():
 
 
 def test_validate_from_ppc_simple_against_target():
-    ppc = get_testgrids('case2_2', 'ppc_testgrids.json')
+    ppc = get_testgrids('ppc_testgrids', 'case2_2.json')
     net = pp.from_json(os.path.join(pp.pp_dir, 'test', 'converter', 'case2_2_by_code.json'))
     assert validate_from_ppc(ppc, net, max_diff_values=max_diff_values1)
 
@@ -88,7 +86,7 @@ def test_ppc_testgrids():
     name = ['case2_1', 'case2_2', 'case2_3', 'case2_4', 'case3_1', 'case3_2', 'case6', 'case14',
             'case57']
     for i in name:
-        ppc = get_testgrids(i, 'ppc_testgrids.json')
+        ppc = get_testgrids('ppc_testgrids', i+'.json')
         net = from_ppc(ppc, f_hz=60)
         validate_other_than_py37(ppc, net, max_diff_values1)
         logger.debug(f'{i} has been checked successfully.')
@@ -100,7 +98,7 @@ def test_pypower_cases():
     name = ['case4gs', 'case6ww', 'case24_ieee_rts', 'case30', 'case39',
             'case118'] # 'case300'
     for i in name:
-        ppc = get_testgrids(i, 'pypower_cases.json')
+        ppc = get_testgrids('pypower_cases', i+'.json')
         net = from_ppc(ppc, f_hz=60)
         validate_other_than_py37(ppc, net, max_diff_values1)
         logger.debug(f'{i} has been checked successfully.')
@@ -108,7 +106,7 @@ def test_pypower_cases():
     # in matpower) another max_diff_values must be used to receive an successful validation
     max_diff_values2 = {"vm_pu": 1e-6, "va_degree": 1e-5, "p_branch_mw": 1e-3,
                         "q_branch_mvar": 1e-3, "p_gen_mw": 1e3, "q_gen_mvar": 1e3}
-    ppc = get_testgrids('case9', 'pypower_cases.json')
+    ppc = get_testgrids('pypower_cases', 'case9.json')
     net = from_ppc(ppc, f_hz=60)
     validate_other_than_py37(ppc, net, max_diff_values2)
 
@@ -148,7 +146,7 @@ def test_to_and_from_ppc():
 
 
 def test_gencost_pwl():
-    case6 = get_testgrids('case6', 'ppc_testgrids.json')
+    case6 = get_testgrids('ppc_testgrids', 'case6.json')
     case6["gencost"] = np.array(
         [
             [1, 0, 0, 2, 1,    0,  5, 7],
@@ -167,7 +165,7 @@ def test_gencost_pwl():
 
 
 def test_gencost_pwl_q():
-    case6 = get_testgrids('case6', 'ppc_testgrids.json')
+    case6 = get_testgrids('ppc_testgrids', 'case6.json')
     case6["gencost"] = np.array(
         [
             [1, 0, 0, 2, 1,    0,  5, 7],
@@ -192,7 +190,7 @@ def test_gencost_pwl_q():
 
 
 def test_gencost_poly_part():
-    case6 = get_testgrids('case6', 'ppc_testgrids.json')
+    case6 = get_testgrids('ppc_testgrids', 'case6.json')
     case6["gencost"] = np.array(
         [
             [2, 0, 0, 2, 14, 0],
@@ -207,7 +205,7 @@ def test_gencost_poly_part():
 
 
 def test_gencost_poly_q():
-    case6 = get_testgrids('case6', 'ppc_testgrids.json')
+    case6 = get_testgrids('ppc_testgrids', 'case6.json')
     case6["gencost"] = np.array(
         [
             [2, 0, 0, 2, 14, 0],
@@ -227,7 +225,7 @@ def test_gencost_poly_q():
 
 
 def test_gencost_poly_q_part():
-    case6 = get_testgrids('case6', 'ppc_testgrids.json')
+    case6 = get_testgrids('ppc_testgrids', 'case6.json')
     case6["gencost"] = np.array(
         [
             [2, 0, 0, 2, 14, 0],
@@ -246,7 +244,7 @@ def test_gencost_poly_q_part():
 
 
 def test_gencost_poly_pwl():
-    case6 = get_testgrids('case6', 'ppc_testgrids.json')
+    case6 = get_testgrids('ppc_testgrids', 'case6.json')
     case6["gencost"] = np.array(
         [
             [1, 0, 0, 3,  1,   0,  5, 7, 10, 10],
@@ -271,7 +269,7 @@ def test_gencost_poly_pwl():
 
 
 def test_gencost_poly_pwl_part_mix():
-    case6 = get_testgrids('case6', 'ppc_testgrids.json')
+    case6 = get_testgrids('ppc_testgrids', 'case6.json')
     case6["gencost"] = np.array(
         [
             [1, 0, 0, 3,  1,   0,  5, 7, 10, 10],